feat: アニメーション付きのメインビジュアルをwipページに反映
