+++ src/app/wip/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { ApplyToProposalSection } from "../../components/ApplyToProposalSection";
 import Footer from "../../components/Footer";
 import { Header } from "../../components/Header";
-import { HeroSection } from "../../components/HeroSection";
+import { HeroSectionWithMotion } from "../../components/HeroSectionWithMotion";
 import { MissionSection } from "../../components/MissionSection";
 import { SponsorsBoardSection } from "../../components/SponsorsBoardSection";
 
@@ -15,7 +15,7 @@ export default function Home() {
     <>
       <Header />
       <main className="pt-16">
-        <HeroSection />
+        <HeroSectionWithMotion />
         <MissionSection />
         <ApplyToProposalSection />
         <SponsorsBoardSection isTopPage={false} />
